Fix not sending correct network properties to connections to other gone

+++ api/Connect/ConnectBridgeToRouterRequest.go
@@ -3,7 +3,7 @@ package api
 type ConnectBridgeToRouterRequest struct {
 	Bridge    string  `json:"bridge"`
 	Router    string  `json:"router"`
-	Latency   int     `json:"latency"`
+	Latency   float64 `json:"latency"`
 	Jitter    float64 `json:"jitter"`
 	DropRate  float64 `json:"dropRate"`
 	Bandwidth int     `json:"bandwidth"`
