+++ api/Connect/ConnectNodeToBridgeRequest.go
@@ -3,7 +3,7 @@ package api
 type ConnectNodeToBridgeRequest struct {
 	Node      string  `json:"node"`
 	Bridge    string  `json:"bridge"`
-	Latency   int     `json:"latency"`
+	Latency   float64 `json:"latency"`
 	Jitter    float64 `json:"jitter"`
 	DropRate  float64 `json:"dropRate"`
 	Bandwidth int     `json:"bandwidth"`
